+++ tests/integration/test_multi.py
@@ -8,20 +8,7 @@ RedisRaft is licensed under the Redis Source Available License (RSAL).
 from pytest import raises
 from redis.exceptions import ExecAbortError, ResponseError
 
-
-class RawConnection(object):
-    """
-    Implement a simple way of executing a Redis command and return the raw
-    unprocessed reply (unlike redis-py's execute_command() which applies some
-    command-specific parsing.)
-    """
-
-    def __init__(self, client):
-        self._conn = client.connection_pool.get_connection('raw-connection')
-
-    def execute(self, *cmd):
-        self._conn.send_command(*cmd)
-        return self._conn.read_response()
+from .sandbox import RawConnection
 
 
 def test_multi_exec_invalid_use(cluster):
